# src/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "chapter 6",
    "contact": {}
  },
  "host": "localhost:3030",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/cars": {
      "get": {
        "summary": "get all cars",
        "tags": [
          "cars admin"
        ],
        "operationId": "getallcars",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "post data cars",
        "tags": [
          "cars admin"
        ],
        "operationId": "postdatacars",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "nama",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "harga",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "ukuran",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "security": [],
...
}